Bugfix for "Foobaz is not an ext filesystem."
I encountered this problem when attempting to install crouton on an
external USB hard drive partition formatted with filesystem ext4.

The issue here is that the if statement is attempting to determine the
filesystem type for the intended chroot installation. It's not evident
to me why it's important that the chroot is installed on an ext[?]
filesystem, but that is what this if statement is checking.

The issue is that when the command df -T "..." is executed it produces
2 lines of output of which the first line needs to be IGNORED. To do
so we insert into the pipe a tail -n 1 to remove the first line.

Now that the output has been reduced to the 1 line of interest, we
want the check in awk to exit with 1 if ^ext is **not** found in
argument 2. To do so, we have to replace the ~ operator with !~.